+++ tests/pdk/finfet_pdk/test_transistor_array.py
@@ -1,3 +1,4 @@
+import os
 import pytest
 from align.pdk.finfet import MOSGenerator
 from .utils import get_test_id, export_to_viewer
@@ -98,6 +99,16 @@ def test_duo_one():
     export_to_viewer(get_test_id(), c)
     if c.drc.num_errors > 0 or len(c.rd.opens) > 0 or len(c.rd.shorts) > 0:
         assert False, f'{get_test_id()}'
+
+
+@pytest.mark.parametrize('vt', ['NMOS', 'PMOS'])
+def test_check_constraints(vt):
+    os.environ['PLACE_ON_GRID'] = 't'
+    c = MOSGenerator()
+    ports = {'S': [('M1', 'S')], 'D': [('M1', 'D')], 'G': [('M1', 'G')]}
+    parameters = {'M': 1, 'NFIN': 4, 'real_inst_type': vt, 'NF': 2}
+    c.addNMOSArray(1, 1, 0, None, ports, **parameters)
+    assert c.metadata["constraints"][0]["constraint"] == "place_on_grid"
 
 
 # Unit tests ###
